fix: 修复golangci-lint警告(noctx, unparam, staticcheck)
- 将exec.Command改为exec.CommandContext以支持context取消
- 修复revive注释格式警告
- 使用time.Until替代Sub(time.Now())
- 移除不必要的类型转换
- 修复未使用参数警告(unparam)

+++ internal/quota/cleanup_unix.go
@@ -11,7 +11,7 @@ import (
 // getFileAccessTime 获取文件访问时间 (Unix).
 func getFileAccessTime(info interface{}) (time.Time, bool) {
 	if stat, ok := info.(*unix.Stat_t); ok {
-		return time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec)), true
+		return time.Unix(stat.Atim.Sec, stat.Atim.Nsec), true
 	}
 	return time.Time{}, false
 }
